Move configuration properties to config package
