+++ src/main/java/uk/gov/hmcts/reform/feature/config/UserConfigurationProperties.java
@@ -1,11 +1,11 @@
-package uk.gov.hmcts.reform.feature.webconsole;
+package uk.gov.hmcts.reform.feature.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 import java.util.List;
 
 @ConfigurationProperties("users")
-public class Ff4jUsersConfig {
+public class UserConfigurationProperties {
 
     private List<UserDetails> admins;
     private List<UserDetails> editors;
